pdf_layouting/requirements.txt
fitz
google-generativeai
pillow